organize footer and split footer into some components
