+++ frontend/src/javascripts/components/molecules/footer-buttons-list.js
@@ -3,8 +3,8 @@ import RevivalBtn from '../atoms/buttons/revival-btn'
 import DeleteButtonsList from '../molecules/delete-buttons-list'
 
 const FooterButtonsList = ({ pathname, rootPath }) => (
-  <div id="disapperance-holder">
-    <ul className="disapperance-list">
+  <div id="action-button-holder">
+    <ul id="action-button-list">
       <RevivalBtn pathname={pathname} rootPath={rootPath} />
       <DeleteButtonsList pathname={pathname} rootPath={rootPath} />
     </ul>
