+++ frontend/src/stylesheets/footer.scss
@@ -77,17 +77,17 @@ $holder-height: 100px;
       width: 400px;
     }
   }
-  #disapperance-holder {
+  #action-button-holder {
     height: $holder-height;
     margin: 0 10px;
     padding: 0;
 
-    ul.disapperance-list {
+    ul#action-button-list {
       margin: 0 5px;
       padding: 0;
       display: flex;
 
-      li.disapperance {
+      li.delete-btn {
         list-style: none;
         width: $holder-height;
         height: $holder-height;
@@ -95,7 +95,7 @@ $holder-height: 100px;
         cursor: pointer;
         display: none;
 
-        .delete-btn {
+        .delete-icon {
           width: $holder-height;
           height: $holder-height;
         }
